backend/access_postgresql/requirements.txt
daphne
Django
python-dotenv
pip
psycopg2-binary
Pillow
python-logstash-async
python-json-logger
channels
channels-redis
dj-database-url
aiosmtpd
djangorestframework
requests
djangorestframework-simplejwt
python-decouple